% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_points.R
\name{clean_points}
\alias{clean_points}
\title{Clean points earned}
\usage{
clean_points(results)
}
\arguments{
\item{results}{Tibble. Feedback or results used for reporting.}
}
\value{
Tibble with clean earned points.
}
\description{
Function applying penalties or partial credits based on question parameters.
}
\author{
Nicolas Mangin
}
