% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_grading.R
\name{compile_grading}
\alias{compile_grading}
\title{Compile all test data}
\usage{
compile_grading(
  test_parameters,
  solutions,
  students,
  closed_answers,
  numeric_answers,
  open_answers,
  open_answers_txt,
  answers,
  parameter_change = NA
)
}
\arguments{
\item{test_parameters}{Tibble.}

\item{solutions}{Tibble.}

\item{students}{Tibble.}

\item{closed_answers}{Tibble.}

\item{numeric_answers}{Tibble.}

\item{open_answers}{Tibble.}

\item{open_answers_txt}{Tibble.}

\item{answers}{Tibble.}

\item{parameter_change}{Character. If there is a change in a question parameters, give the question id.}
}
\value{
List with all compiled and updated test data.
}
\description{
Function Gathering and updating test data.
}
\author{
Nicolas Mangin
}
