% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading_server.R
\name{grading_server}
\alias{grading_server}
\title{Grade tests.}
\usage{
grading_server(id, test, tree, course_data, course_paths)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{test}{Reactive. Selected test.}

\item{tree}{Reactive. Function containing a list of documents as a classification tree compatible with jsTreeR.}

\item{course_data}{Reactive. Function containing all the course data loaded with the course.}

\item{course_paths}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}
}
\value{
Save the test results in the relevant test sub-folder.
}
\description{
Module allowing the user to grade, check, and calibrate tests.
}
\author{
Nicolas Mangin
}
