% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_scale.R
\name{make_scale}
\alias{make_scale}
\title{Create a grading scale}
\usage{
make_scale(letter, proposition, scale, checked, correct, ns)
}
\arguments{
\item{letter}{Character. Letter in the propositions database.}

\item{proposition}{Character. Answer/criterion selected or given.}

\item{scale}{Character. Whether the initial scale of the letter is "logical", "numeric", a "percentage", or "qualitative".}

\item{checked}{Integer. 1 if the student provided this answer, 0 otherwise.}

\item{correct}{Numeric. Whether the answer is correct.}

\item{ns}{Function. Function assigning input IDs to the relevant module.}
}
\value{
A shiny user interface for checking or grading answers.
}
\description{
Function creating the user interface to grade an item.
}
\author{
Nicolas Mangin
}
