% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinitialize_weights.R
\name{reinitialize_weights}
\alias{reinitialize_weights}
\title{Re-initialize weights}
\usage{
reinitialize_weights(test_parameters, feedbacks, selected_question = NULL)
}
\arguments{
\item{test_parameters}{Tibble. Test parameters.}

\item{feedbacks}{Tibble. Local feedback files combined.}

\item{selected_question}{Character. Question file if one question in particular is targeted.}
}
\value{
Feedback with updated weights
}
\description{
Function computing new weights for items of each question based on test parameters.
}
\author{
Nicolas Mangin
}
