% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_answers.R
\name{update_answers}
\alias{update_answers}
\title{Create an answer file.}
\usage{
update_answers(test_name, scoring, answers = NULL, input = NULL)
}
\arguments{
\item{test_name}{Character. Name of the test.}

\item{scoring}{Tibble. List of weighted letters and propositions.}

\item{answers}{Tibble. Tables including all information necessary to compute weights.}

\item{input}{List. List containing the checks entered by the use when manually grading.}
}
\value{
A tibble which can be edited, updated and saved containing students' answers before scoring.
}
\description{
Function creating the answer file used for grading.
}
\author{
Nicolas Mangin
}
