% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_propositions.R
\name{update_propositions}
\alias{update_propositions}
\title{Update feedbacks}
\usage{
update_propositions(
  test_name,
  test_parameters = NULL,
  feedbacks = NULL,
  edits = NULL
)
}
\arguments{
\item{test_name}{Character. Name of the test.}

\item{test_parameters}{Tibble. Test parameters.}

\item{feedbacks}{Tibble. Local feedback files combined.}

\item{edits}{Tibble. Changes made to the feedbacks (from interface)}
}
\value{
Tibble. New feedback saved in local feedback files (one for each version of each question).
}
\description{
Function updating locally feedbacks based on user input.
}
\author{
Nicolas Mangin
}
