% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_question.R
\name{update_question}
\alias{update_question}
\title{Update question parameters}
\usage{
update_question(
  test_name = NULL,
  test_parameters = NULL,
  edited_question = NULL,
  new_points = NULL,
  new_penalty = NULL,
  new_partial_credits = NULL
)
}
\arguments{
\item{test_name}{Character. Name of the test.}

\item{test_parameters}{Tibble. Items checked in answers.}

\item{edited_question}{Character. Code of the question to edit.}

\item{new_points}{Numeric. Points assigned to the question.}

\item{new_penalty}{Logical. Whether negative points should be counted.}

\item{new_partial_credits}{Logical Whether partial credits should be earned.}
}
\value{
Tibble. Updated test parameters.
}
\description{
Function updating question parameters (penalty, partial credits, points) based on user input.
}
\author{
Nicolas Mangin
}
