% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_results.R
\name{update_results}
\alias{update_results}
\title{Update results}
\usage{
update_results(test_name, scoring, answers)
}
\arguments{
\item{test_name}{Character. Name of the test.}

\item{scoring}{Tibble. Table containing the weights for each question.}

\item{answers}{Tibble. Table containing the checks for each student on each question.}
}
\value{
Tibble with the points students earned on each
}
\description{
Function gathering choices from test_parameters, checks from answers, and weights from feedbacks to produce the results file.
}
\author{
Nicolas Mangin
}
