% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_scoring.R
\name{update_scoring}
\alias{update_scoring}
\title{Update results}
\usage{
update_scoring(test_parameters, feedbacks)
}
\arguments{
\item{test_parameters}{Tibble. Test parameters defined when the test was created (or adjusted subsequently)}

\item{feedbacks}{Tibble. Local feedback files combined.}
}
\value{
Tibble with the points students earned on each
}
\description{
Function gathering choices from test_parameters, checks from answers, and weights from feedbacks to produce the results file.
}
\author{
Nicolas Mangin
}
