% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading_ui.R
\name{grading_ui}
\alias{grading_ui}
\title{Grade tests.}
\usage{
grading_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
Save the test results in the relevant test sub-folder.
}
\description{
Module allowing the user to grade, check, and calibrate tests.
}
\author{
Nicolas Mangin
}
